refactored FiltersPanel
